Reorganize pulse visualization code (#2340)
An improved pulse visualizer was added in #2213, but in that commit the
additional matplotlib logic and utility functions were added on to
existing modules that used matplotlib. This makes maintaining the code
more difficult for 2 reasons, first the pulse visualization code is
mostly self contained and didn't really have anything to with the
modules it was added onto, and secondly the existing matplotlib code
was already quite large. This commit reorganizes the pulse visualization
code to be easier to maintain by moving all pulse related lib code into
separate modules in a pulse/ directory.
