+++ qiskit/pulse/commands/sample_pulse.py
@@ -56,7 +56,7 @@ class SamplePulse(Command):
 
         Args:
             dt: Time interval of samples.
-            style (OPStylePulse): A style sheet to configure plot appearance
+            style (PulseStyle): A style sheet to configure plot appearance
             filename: Name required to save pulse image
             interp_method: A function for interpolation
             scaling (float): Relative visual scaling of waveform amplitudes
@@ -68,7 +68,7 @@ class SamplePulse(Command):
         """
         # pylint: disable=invalid-name, cyclic-import
 
-        from qiskit.tools import visualization
+        from qiskit import visualization
 
         return visualization.pulse_drawer(self, dt=dt, style=style, filename=filename,
                                           interp_method=interp_method, scaling=scaling,
