+++ qiskit/pulse/schedule.py
@@ -196,7 +196,7 @@ class Schedule(ScheduleComponent):
 
         Args:
             dt: Time interval of samples
-            style (OPStyleSched): A style sheet to configure plot appearance
+            style (SchedStyle): A style sheet to configure plot appearance
             filename: Name required to save pulse image
             interp_method: A function for interpolation
             scaling (float): Relative visual scaling of waveform amplitudes
@@ -214,7 +214,7 @@ class Schedule(ScheduleComponent):
         """
         # pylint: disable=invalid-name, cyclic-import
 
-        from qiskit.tools import visualization
+        from qiskit import visualization
 
         return visualization.pulse_drawer(self, dt=dt, style=style,
                                           filename=filename, interp_method=interp_method,
