+++ qiskit/visualization/pulse/interpolation.py
