+++ qiskit/visualization/pulse_visualization.py
@@ -20,7 +20,7 @@ matplotlib pulse visualization.
 
 from qiskit.pulse import Schedule, Instruction, SamplePulse
 from qiskit.visualization.exceptions import VisualizationError
-from qiskit.visualization import matplotlib as _matplotlib
+from qiskit.visualization.pulse import matplotlib as _matplotlib
 
 
 def pulse_drawer(data, dt=1, style=None, filename=None,
@@ -32,7 +32,8 @@ def pulse_drawer(data, dt=1, style=None, filename=None,
     Args:
         data (ScheduleComponent or SamplePulse): Data to plot
         dt (float): Time interval of samples
-        style (OPStylePulse or OPStyleSched): A style sheet to configure plot appearance
+        style (PulseStyle or SchedStyle): A style sheet to configure
+            plot appearance
         filename (str): Name required to save pulse image
         interp_method (Callable): interpolation function
             See `qiskit.visualization.interpolation` for more information
